Encapsulate quick_sort into value_sort module - fix quick_sort borrowing issues
